+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.2.42
 - fixing tardis flight and crash data
+- sonic can see repair time left